+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Upcoming release
 
+### Features
+
+- new rule `heading-level` for validating sequential heading levels.
+
+### Bugfixes
+
+- `Table.getMetaFor(..)` is not case-insensitive.
+
 ## 0.10.0 (2018-08-11)
 
 ### Breaking changes